Player ID is now an int, not a long
It is no longer the same as the PeerID (which is a long)
Shots, Creatures and Players each have their own ID pool for no particular reason

+++ lotshared/Snapshot.cs
@@ -53,7 +53,7 @@ namespace lotshared
             }
             for (var i = 0; i < pCount; i++)
             {
-                Player p = new Player(reader.GetLong());
+                Player p = new Player(reader.GetInt());
                 p.X = reader.GetInt();
                 p.Y = reader.GetInt();
                 snapshot.Players.Add(p.Id, p);
